+++ final/FinalProject/DC_Voltage.cs
@@ -1,7 +1,7 @@
 public class DC_Voltage : Voltage
 {
     // Constructor
-    public DC_Voltage(string name, int voltage) : base(name, voltage)
+    public DC_Voltage(string name, float voltage) : base(name, voltage)
     {
         _elementType = "DC Voltage Source";
         _phaseAngle = 0;
